tests/integration: Move test "extends_w_file_subdir" to corresp. dir
