+++ tests/integration/extends_w_file_subdir/__init__.py
@@ -1,0 +1,1 @@
+
